refactored all home screen buttons, started refactoring trip list screen
