+++ screens/LoginForm.js
@@ -120,11 +120,14 @@ const styles = StyleSheet.create({
   loginInput: {
     fontSize: 18,
     paddingVertical: 10,
+    paddingLeft: 5,
     width: 250,
     margin: 4,
     alignItems: "flex-start",
     fontWeight: "normal",
     fontFamily: "Futura-Medium",
+    backgroundColor: "#f0f4f7",
+    borderRadius: 5
   },
   closeButton: {
     borderRadius: 5,
